move system info functions into a separate module
so it can be shared by variant_dir and install_root
